Fix API for non-Typescript clients
Fixes microsoft/vscode-textbuffer#2

Includes test code that demonstrates the issue.

Typescript const enums are not exported in a way accessible
to non-Typescript clients. This fixes the issue by simply
utilizing the `preserveConstEnums` compiler option. This may
have a performance impact; If that is unacceptable there are
other ways to fix this issue, but the included test file is
useful to keep going forward.